feat: enhance footer links with icons and restructure IFooterLinks interface
